feat: separate destination test into two part
